Add nil-guarding for nilaway linter

+++ account.go
@@ -114,6 +114,9 @@ func OpenAccount(rootPath, passphrase string) (*Account, error) {
 	if err != nil {
 		return nil, err
 	}
+	if decryptedFile == nil {
+		return nil, errors.New("openpgp.ReadMessage returned nil")
+	}
 
 	entity, err := openpgp.ReadEntity(packet.NewReader(decryptedFile.UnverifiedBody))
 	if err != nil {
@@ -132,6 +135,9 @@ type Account struct {
 }
 
 func (a *Account) Identity() (string, error) {
+	if a == nil || a.entity == nil {
+		return "", ErrNoIdentity
+	}
 	for _, i := range a.entity.Identities {
 		return i.Name, nil
 	}
@@ -140,6 +146,9 @@ func (a *Account) Identity() (string, error) {
 }
 
 func (a *Account) Name() string {
+	if a == nil || a.entity == nil {
+		return ""
+	}
 	for _, i := range a.entity.Identities {
 		return i.UserId.Name
 	}
@@ -148,6 +157,9 @@ func (a *Account) Name() string {
 }
 
 func (a *Account) Email() string {
+	if a == nil || a.entity == nil {
+		return ""
+	}
 	for _, i := range a.entity.Identities {
 		return i.UserId.Email
 	}
@@ -156,10 +168,16 @@ func (a *Account) Email() string {
 }
 
 func (a *Account) Fingerprint() Fingerprint {
+	if a == nil || a.entity == nil || a.entity.PrimaryKey == nil {
+		return Fingerprint{}
+	}
 	return a.entity.PrimaryKey.Fingerprint
 }
 
 func (a *Account) Export(w io.Writer) error {
+	if a == nil || a.entity == nil {
+		return errors.New("account or entity is nil")
+	}
 	armored, err := armor.Encode(w, openpgp.PublicKeyType, map[string]string{})
 	if err != nil {
 		return err
@@ -175,6 +193,10 @@ func (a *Account) Export(w io.Writer) error {
 }
 
 func (a *Account) certificate(DNSNames []string) (cert tls.Certificate, err error) {
+	if a == nil || a.entity == nil || a.entity.PrimaryKey == nil || a.entity.PrivateKey == nil {
+		err = errors.New("account or entity is incomplete")
+		return
+	}
 	template := x509.Certificate{
 		Version:      3,
 		DNSNames:     DNSNames,
@@ -288,6 +310,9 @@ func (a *Account) AddFile(r io.Reader, name string, recipients []*Friend) (*File
 	if err != nil {
 		return nil, err
 	}
+	if w == nil {
+		return nil, errors.New("openpgp.Encrypt returned nil writer")
+	}
 
 	if _, err := io.Copy(w, r); err != nil {
 		return nil, err
@@ -298,6 +323,10 @@ func (a *Account) AddFile(r io.Reader, name string, recipients []*Friend) (*File
 }
 
 func (a *Account) RemoveFile(file *File) error {
+	if file == nil {
+		return nil
+	}
+	
 	err := os.Remove(file.Path)
 	if err != nil {
 		return err
